move StatisticsUtils tests into separate class
